Revert "update to unicode 11"
