+++ .gitignore
@@ -1,5 +1,3 @@
 uniemoji.db
-uniemoji.xml
-config.py
 *~
 *.pyc
