+++ HISTORY
@@ -1,20 +1,7 @@
-# 0.6.0 🐍🐍🐍📁 — 2016-10-15
+# Next release: WIP
 
-- UniEmoji now requires Python 3
-- The IBus window finally responds to mouse clicks (thanks Mike Fabian @mike-fabian)
-- Currency symbols added to the list of candidates (Mike Fabian)
-- The installation can now be customized with the variables PREFIX, SYSCONFDIR,
-  DATADIR, DESTDIR, PYTHON (thanks Brian Tarricone @kelnos, Roland Hieber @rohieb, Kirill Boltaev @aske)
-
-# 0.5.0 🤜🤛 — 2016-06-22
-
-- Update UnicodeData to 9.0
-- Input using well-known “shortnames”, e.g. :​grin: or :​eggplant:.
-  The table of shortnames comes courtesy of [EmojiOne](http://emojione.com/).
-  Various EmojiOne aliases are also supported.
-- Allow searching using punctuation. Useful for aliases such as ':)'
-- Dashes have been added to the list of candidates (en-dash, em-dash, etc.)
-- Project transferred to a new maintainer
+- Input using well-known “shortnames”, e.g. :grin: or :+1:. The table
+  of shortnames comes courtesy of [Emoji One](http://emojione.com/).
 
 # 0.4.0 👦🏽🌮 — 2015-10-24
 
@@ -47,4 +34,4 @@
 
 # 0.1.0: 🐱🚀 — 2013-01-14
 
-- Initial release
+- Intiial release
