--- config.py.in
@@ -1,1 +1,0 @@
-SYS_CONF_DIR = '@SYSCONFDIR@'
