+++ custom.json
@@ -1,6 +1,4 @@
 {
-  "shruggie": "¯\\_(ツ)_/¯",
-  "kyubei": "／人◕ ‿‿ ◕人＼",
-  "lenny":"( ͡° ͜ʖ ͡°)",
-  "look of disapproval":"ಠ_ಠ"
+  "shrug": "¯\\_(ツ)_/¯",
+  "kyubei": "／人◕ ‿‿ ◕人＼"
 }
