+++ uniemoji.xml
@@ -7,7 +7,7 @@
     <license>GPL</license>
     <author>Lalo Martins &lt;lalo.martins@gmail.com&gt;</author>
     <homepage>https://github.com/salty-horse/ibus-uniemoji</homepage>
-    <exec>@PYTHON@ @DATADIR@/ibus-uniemoji/ibus.py --ibus</exec>
+    <exec>/usr/bin/python2 /usr/share/ibus-uniemoji/uniemoji.py --ibus</exec>
     <textdomain>uniemoji</textdomain>
     <engines>
         <engine>
@@ -17,12 +17,12 @@
             <language></language>
             <license>GPL</license>
             <author>Lalo Martins &lt;lalo.martins@gmail.com&gt;</author>
-            <icon>@DATADIR@/ibus-uniemoji/uniemoji.svg</icon>
-            <layout>us</layout>
-            <layout_variant></layout_variant>
-            <layout_option></layout_option>
+            <icon>/usr/share/ibus-uniemoji/uniemoji.svg</icon>
+            <layout>default</layout>
+            <layout_variant>default</layout_variant>
+            <layout_option>default</layout_option>
             <hotkeys></hotkeys>
-            <symbol>🐱</symbol>
+            <symbol></symbol>
             <setup></setup>
             <version></version>
             <textdomain></textdomain>
